+++ core/src/main/java/com/fpf/smartscansdk/core/processors/BatchProcessor.kt
@@ -10,10 +10,9 @@ import kotlinx.coroutines.sync.withPermit
 import kotlinx.coroutines.withContext
 import java.util.concurrent.atomic.AtomicInteger
 
-// For BatchProcessor’s use case—long-running, batched,  asynchronous processing—the Application context should be used.
 abstract class BatchProcessor<Input, Output>(
     private val context: Context,
-    protected val listener: ProcessorListener<Input, Output>? = null,
+    protected val listener: ProcessorListener<Input>? = null,
     private val memoryOptions: MemoryOptions = MemoryOptions(),
     val batchSize: Int = 10
 ) {
@@ -21,7 +20,7 @@ abstract class BatchProcessor<Input, Output>(
         const val TAG = "BatchProcessor"
     }
 
-    open suspend fun run(items: List<Input>): Metrics = withContext(Dispatchers.IO) {
+    suspend fun run(items: List<Input>): ProcessorResult = withContext(Dispatchers.IO) {
         val processedCount = AtomicInteger(0)
         val startTime = System.currentTimeMillis()
         var totalSuccess = 0
@@ -29,9 +28,9 @@ abstract class BatchProcessor<Input, Output>(
         try {
             if (items.isEmpty()) {
                 Log.w(TAG, "No items to process.")
-                val metrics = Metrics.Success()
-                listener?.onComplete(context.applicationContext, metrics)
-                return@withContext metrics
+                val processorResult = ProcessorResult.Success()
+                listener?.onComplete(context.applicationContext, processorResult)
+                return@withContext processorResult
             }
 
             val memoryUtils = Memory(context.applicationContext, memoryOptions)
@@ -47,10 +46,9 @@ abstract class BatchProcessor<Input, Output>(
                         semaphore.withPermit {
                             try {
                                 val output = onProcess(context.applicationContext, item)
-                                output
+                                item to Result.success(output)
                             } catch (e: Exception) {
-                                listener?.onError(context.applicationContext, e, item)
-                                null
+                                item to Result.failure(e)
                             }finally {
                                 val current = processedCount.incrementAndGet()
                                 val progress = current.toFloat() / items.size
@@ -60,28 +58,48 @@ abstract class BatchProcessor<Input, Output>(
                     }
                 }
 
-                val outputBatch = deferredResults.mapNotNull { it.await() }
-                totalSuccess += outputBatch.size
-                onBatchComplete(context.applicationContext, outputBatch)
+                val successfulResults = mutableListOf<Output>()
+
+                try {
+                    for (deferred in deferredResults) {
+                        val result = deferred.await()
+                        if (result.second.isSuccess) {
+                            successfulResults += result.second.getOrThrow()
+                        } else {
+                            val error = result.second.exceptionOrNull() as Exception
+                            listener?.onError(context.applicationContext, error, result.first)
+                        }
+                    }
+                } catch (e: Exception) {
+                    // Fatal item error, cancel all remaining in batch
+                    deferredResults.forEach {
+                        it.cancel()
+                    }
+                    throw e
+                }
+
+                totalSuccess += successfulResults.size
+                onBatchComplete(context.applicationContext, successfulResults)
             }
 
             val endTime = System.currentTimeMillis()
-            val metrics = Metrics.Success(totalSuccess, timeElapsed = endTime - startTime)
+            val processorResult = ProcessorResult.Success(totalSuccess, timeElapsed = endTime - startTime)
 
-            listener?.onComplete(context.applicationContext, metrics)
-            metrics
+            listener?.onComplete(context.applicationContext, processorResult)
+            processorResult
         }
         catch (e: CancellationException) {
+            listener?.onCancel(context.applicationContext)
             throw e
         }
         catch (e: Exception) {
-            val metrics = Metrics.Failure(
+            val processorResult = ProcessorResult.Failure(
                 totalProcessed = totalSuccess,
                 timeElapsed = System.currentTimeMillis() - startTime,
                 error = e
             )
-            listener?.onFail(context.applicationContext, metrics)
-            metrics
+            listener?.onFail(context.applicationContext, processorResult)
+            processorResult
         }
     }
 
@@ -94,8 +112,4 @@ abstract class BatchProcessor<Input, Output>(
     protected abstract suspend fun onBatchComplete(context: Context, batch: List<Output>)
 
 }
-
-
-
-